requirements.txt
fastapi[all]
uvicorn[standard]
aiosqlite
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
argon2-cffi